requirements.txt
asgiref==3.8.1
boto3==1.35.34
botocore==1.35.34
certifi==2024.8.30
charset-normalizer==3.3.2
crispy-bootstrap5==2024.10
dj-database-url==2.2.0
Django==5.1.1
django-allauth==65.0.2
django-countries==7.6.1
django-crispy-forms==2.3
django-storages==1.14.4
djangorestframework==3.15.2
gunicorn==23.0.0
idna==3.10
jmespath==1.0.1
packaging==24.1
pillow==10.4.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
s3transfer==0.10.2
six==1.16.0
sqlparse==0.5.1
stripe==11.1.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
